+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - **Breaking:** The minimal required Python release is 3.6.
 - **Breaking:** rohrpost has now an explicit dependency on Channels >= 3.
 - Make support for Python 3.9 official.
+- ``TolerantJSONEncoder.default()`` no properly delegates to ``JSONEncoder.default()``
 
 2.0.1
 -----